`RegisterManager`'s `Size` method now returns `int` type

+++ usm64/managers/registers.go
@@ -25,8 +25,8 @@ func (m *RegisterMap) DeleteRegister(reg *gen.RegisterInfo) core.ResultList {
 	return core.ResultList{}
 }
 
-func (m *RegisterMap) Size() uint {
-	return uint(len(*m))
+func (m *RegisterMap) Size() int {
+	return len(*m)
 }
 
 func (m *RegisterMap) GetAllRegisters() []*gen.RegisterInfo {
